refractor: seperate user module into profile & new user module
