+++ backend/src/profile/profile.module.ts
@@ -1,0 +1,12 @@
+import { Module } from '@nestjs/common';
+import { ProfileController } from './profile.controller';
+import { ProfileService } from './profile.service';
+import { CommonModule } from '@common/common.module';
+
+@Module({
+  imports: [CommonModule],
+  controllers: [ProfileController],
+  providers: [ProfileService],
+  exports: [ProfileService],
+})
+export class ProfileModule {}
